V 0.4
 *WIP*
-Profile
-Stats Tracking
-Support Features
-Friend code auto send feature

*ADDED*
-Security Features
-Welcome/Introduction
-Can only use the server when you give userself ranks

*FINISHED*
-Auto Channels
-Stats tracking channel now complete
-MM-Search-Voice Channel with rank checking

*BUGFIXES*
A BUG WHERE:
NONE

+++ src/de/mrcloud/utils/JDAUtils.java
@@ -2,6 +2,9 @@ package de.mrcloud.utils;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.*;
+import net.dv8tion.jda.api.events.Event;
+import net.dv8tion.jda.api.events.GenericEvent;
+import net.dv8tion.jda.api.events.guild.GenericGuildEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 
 import java.awt.*;
@@ -24,33 +27,45 @@ public class JDAUtils {
     }
 
 
-    public void RedBuilder(String Title, GuildMessageReceivedEvent e, TextChannel txtChannel, String ErrorText, int deleteAfter) {
+    public void RedBuilder(String Title, Member member, TextChannel txtChannel, String ErrorText, int deleteAfter, boolean delete) {
         EmbedBuilder embBuilder = new EmbedBuilder();
         embBuilder.setTitle(Title);
-        embBuilder.setAuthor(e.getAuthor().getName(), e.getAuthor().getAvatarUrl(), e.getAuthor().getAvatarUrl());
+        embBuilder.setAuthor(member.getUser().getName(), member.getUser().getAvatarUrl(), member.getUser().getAvatarUrl());
         embBuilder.setColor(Color.decode("#d63031"));
         embBuilder.setDescription(ErrorText);
-        txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        if (delete) {
+            txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        } else {
+            txtChannel.sendMessage(embBuilder.build()).queue();
+        }
 
     }
 
-    public void YellowBuilder (String Title,GuildMessageReceivedEvent e, TextChannel txtChannel, String InfoText, int deleteAfter) {
+    public void YellowBuilder(String Title, Member member, TextChannel txtChannel, String InfoText, int deleteAfter, boolean delete) {
         EmbedBuilder embBuilder = new EmbedBuilder();
         embBuilder.setTitle(Title);
-        embBuilder.setAuthor(e.getAuthor().getName(), e.getAuthor().getAvatarUrl(), e.getAuthor().getAvatarUrl());
+        embBuilder.setAuthor(member.getUser().getName(), member.getUser().getAvatarUrl(), member.getUser().getAvatarUrl());
         embBuilder.setColor(Color.decode("#f9ca24"));
         embBuilder.setDescription(InfoText);
-        txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        if (delete) {
+            txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        } else {
+            txtChannel.sendMessage(embBuilder.build()).queue();
+        }
 
     }
 
-    public void GreenBuilder(GuildMessageReceivedEvent e, String Title, TextChannel txtChannel, String InfoText, int deleteAfter) {
+    public void GreenBuilder(String Title, Member member, TextChannel txtChannel, String InfoText, int deleteAfter, boolean delete) {
         EmbedBuilder embBuilder = new EmbedBuilder();
         embBuilder.setTitle(Title);
-        embBuilder.setAuthor(e.getAuthor().getName(), e.getAuthor().getAvatarUrl(), e.getAuthor().getAvatarUrl());
+        embBuilder.setAuthor(member.getUser().getName(), member.getUser().getAvatarUrl(), member.getUser().getAvatarUrl());
         embBuilder.setColor(Color.decode("#2ecc71"));
         embBuilder.setDescription(InfoText);
-        txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        if (delete) {
+            txtChannel.sendMessage(embBuilder.build()).complete().delete().queueAfter(deleteAfter, TimeUnit.SECONDS);
+        } else {
+            txtChannel.sendMessage(embBuilder.build()).queue();
+        }
 
     }
 
@@ -90,5 +105,11 @@ public class JDAUtils {
         embBuilder.setColor(Color.decode("#2ecc71"));
         embBuilder.setDescription("Downloaded File " + e.getMessage().getAttachments().get(0).getFileName());
         txtChannel.sendMessage(embBuilder.build()).complete();
+
+    }
+
+    public List<Member> getTeamMembers(Guild server, String roleName) {
+        return server.getMembersWithRoles(server.getRolesByName(roleName, true).get(0));
     }
 }
+
